refactor(persona): extract _LLMPersonaAgent to persona_runtime.py (RFC 0005, PR 8d)
Extract _LLMPersonaAgent, helper functions (_truncate_with_ellipsis,
_coerce_event_timeout), and 8 private constants from persona.py into
new agents/persona_runtime.py module.

persona.py shrinks from 1076 to 256 lines (ABC + factory + re-exports).
persona_runtime.py contains the concrete LLM-powered decision loop.

Circular import resolved via late import: persona_runtime.py imports
PersonaAgent at top level; persona.py imports _LLMPersonaAgent after
PersonaAgent class definition.

Changes:
- NEW: agents/persona_runtime.py (869 lines)
- agents/persona.py: removed extracted code, added late import re-exports
- agents/server.py: import _LLMPersonaAgent from persona_runtime
- agents/dispatch.py: TYPE_CHECKING import from persona_runtime
- agents/tick.py: TYPE_CHECKING import from persona_runtime
- tests: updated imports, added persona_runtime to import smoke tests
  and circular import isolation test, added backward-compat re-export
  test, added metadata deep-copy isolation test (F-64-R-SF3)
- ROADMAP.md: added missing PR #64 (8a) entry, updated component table
- docs/rfcs/0005-pr-plan.md: checked PR 8d checklist items

234 tests pass, ruff clean.
