+++ FILEMAP.md
@@ -1,14 +1,14 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-14  
-> **187** tracked files
+> **188** tracked files
 
 ## Summary
 
 | Top-level directory | Files |
 |---------------------|-------|
 | `internal/` | 40 |
-| `agents/` | 32 |
+| `agents/` | 33 |
 | `tests/` | 28 |
 | `docs/` | 21 |
 | `(root)` | 19 |
@@ -26,7 +26,7 @@
 
 | Extension | Count |
 |-----------|-------|
-| `.py` | 69 |
+| `.py` | 70 |
 | `.md` | 35 |
 | `.go` | 35 |
 | `.yaml` | 22 |
@@ -95,6 +95,7 @@
 │   │   └── working.py
 │   ├── persona.py
 │   ├── persona_behavior.py
+│   ├── persona_runtime.py
 │   ├── persona_types.py
 │   ├── pyproject.toml
 │   ├── server.py
