+++ agents/server.py
@@ -26,7 +26,8 @@ from .base import BaseAgent, TaskInput, TaskInputConfig, TaskOutput, TaskStatus
 from .dispatch import EventDispatcher
 from .generated import task_pb2, task_pb2_grpc
 from .llm_client import LLMClient, create_provider
-from .persona import _LLMPersonaAgent, create_persona_agent
+from .persona import create_persona_agent
+from .persona_runtime import _LLMPersonaAgent
 from .task_agent import TaskAgent
 from .tick import TickScheduler
 from .tools import builtin
