+++ tests/integration/test_persona_e2e.py
@@ -15,7 +15,8 @@ import pytest
 from agents.generated import task_pb2, task_pb2_grpc
 from agents.llm_client import LLMClient, LLMResponse, StopReason, Usage
 from agents.dispatch import ActionExecutor, EventDispatcher
-from agents.persona import _LLMPersonaAgent, create_persona_agent
+from agents.persona import create_persona_agent
+from agents.persona_runtime import _LLMPersonaAgent
 from agents.persona_types import ActionType, AgentEvent, EventType
 from agents.tick import TickScheduler
 from agents.server import AgentServiceServicer
